Extract Paths, Depths, and Values from a Nested List

This function recursively traverses a nested list in R and returns the paths, depths, and values of each element. Optionally, you can filter the results to only include elements with a specific name. Unnamed elements are represented by a special character "_".

extractPathsDepthsAndValues(
  nestedList,
  currentPath = "",
  depth = 1,
  item = NULL
)

Arguments

nestedList

A nested list object from which to extract paths, depths, and values.

currentPath

A character string representing the current path during recursion. Defaults to an empty string. This parameter is primarily used internally by the function.

depth

An integer representing the current depth of the recursion. Defaults to 1. This parameter is primarily used internally by the function.

item

An optional character string specifying the name of the item to look for. If provided, the function will return only the paths, depths, and values for elements matching this name. If NULL, all elements are included. Defaults to NULL.

Value

A data frame with three columns: "path", "depth", and "value", where each row represents an element in the nested list.